backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.19
pandas==2.1.3
numpy==1.26.2
pandas-ta==0.3.14b
sqlalchemy==2.0.23
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.1
